requirements.txt
pandas
openai
yagmail
python-dotenv
customtkinter
langchain
langchain-openai
openpyxl
streamlit
